Use common constant for depth of coverage DP, overridable via XX flag. Default DP is now 2 (from 0, 2 or 4, depending on where you looked).

+++ src/com/rtg/variant/coverage/CoverageProcessor.java
@@ -61,6 +61,6 @@ public abstract class CoverageProcessor implements Closeable {
    * @param coverage coverage for the region
    * @throws IOException because it might do some output
    */
-  public void finalCoverageRegion(String name, int start, int end, int coverage) throws IOException { }
+  public void finalCoverageRegion(String name, int start, int end, double coverage) throws IOException { }
 
 }
